ref(symcache): Move transformer tests out to a separate file
